[IMP] l10n_mx_edi_stock{_extended}_{30,31}: add Carta Porte 3.0 PDF
[This commit] added in a new Carta Porte 3.0 PDF for Mexico in version
saas-17.3 that inherits from the standard delivery slip, by removing
all the tables related to the products etc. (for security reasons: the
driver doesn't need to know what's in the cargo) and adding in some new
required fields.

This commit backports the report from saas-17.3 with the following
changes:
- We improved the layout a bit to make the report even clearer.
- We corrected the "Num Reg ID Trib" and "Residencia Fiscal" of the
  "Figura"'s to use the right value and show them per figura.
- We corrected the departure and arrival date (they were swapped).

These changes will be forward-ported again to the upper versions.

[This commit]: https://github.com/odoo/enterprise/commit/af14612ec5942acdc06584ec8ff9d8bf26998ae7

task-3990042

closes odoo/enterprise#68059

X-original-commit: 1f0f2c0802724d85d06e98a1d00085ce8f252156
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>