[FIX] l10n_pe_reports: fix tests
With this commit: https://github.com/odoo/odoo/pull/166896/commits/90088bd97ec096fa8c360868e3a1addef91bd4ce

We automatically add the country of the chart template to the company if the
company does not have one by default. This change had a side effect that in the
context of the test "TestPeSales," the boolean field l10n_latam_use_documents
on the journal was now set to True.

The consequence of that was that the constraint _check_l10n_latam_documents was
breaking since this constraint checks that for the posted moves that use
documents, the move must have a document type and a document number.

closes odoo/enterprise#64991

Task: 3945833
Related: odoo/odoo#166896
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>