[IMP] *: adapt to composer's changes
* = industry_fsm_report, test_l10n_be_hr_payroll_account

Purpose:
--------
In the community counterpart, the "send" button of the mail composer is
replaced by a dropdown button when the composer is in mono-comment mode,
to add an option that allows users to schedule the posting of messages.

This commit adapts selectors in tours accordingly.

Task-3877073

closes odoo/enterprise#67957

Related: odoo/odoo#175683
Related: odoo/upgrade#6344
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>