[FIX] l10n_in_qr_code_bill_scan: resolve props validation traceback
Steps to reproduce
==================
1. Go to the Accounting module
2. Open Vendor Bills
3. Click on the Scan button with debugger mode ON

=> Hence gives traceback

Cause of the issue
==================
The Issue was due to the `BillQrScan` component doing the props validation
same as the `Dialog` component, the `slots` object is a mandatory prop for Dialog
component but not needed for `BillQrScan`.

This commit solves the traceback generating due to props validation on
`BillQrScan` component.

closes odoo/enterprise#65335

X-original-commit: fce746701024e14644cebba8ecd2d85a42c85a8a
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Vivek Pathak (vivp) <vivp@odoo.com>