[FIX] spreadsheet_edition,documents_spreadsheet: move scss to correct module
Steps to reproduces:
- install module `test_spreadsheet_edition`
- run js tests
=> some tests fail.

The reason is we are trying to click or hover some cells in those tests.
But the spreadsheet (its viewport) size is 0.
The cells are not visible.

This commit moves some scss from module `documents_spreadsheet` to
`spreadsheet_edition`.

This css ensures the spreadsheet takes all the available space and be sized properly.

runbot errors 65775, 65774, 65773, 65772

closes odoo/enterprise#65123

X-original-commit: 9236a035b8d32c1bf1220f59b462e80a19341b5a
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>