[IMP] website_helpdesk_knowledge: enable selection of any published article
In the helpdesk settings, the admin can define the article the user will
land on when clicking on the "Browse articles" button from the website
help page. The selector currently excludes the articles having a parent.

To enable the selection of any published article, we will remove the
condition `('parent_id', '=', False)` from the search domain.

task-3980714

X-original-commit: daaa0521a2be438af241344f4670c1d00f1d412f
Part-of: odoo/enterprise#68360
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>