[FIX] : correct drag and drop sign box coordinates
To reproduce the bug, follow these steps:
- Download the document present in ticket 3991731.
- Upload it to the sign app.
- Try to drag and drop the sign box to the bottom.

You will find that the drag and drop functionality in general is erratic. This is due to the coordinates being incorrect, sometimes even negative, since the method used to calculate the height and width of the PDF view is flawed .

I used the getBoundingClientRect method (the same method used to get the sign box coordinates) to fix this issue.

opw-3991731

closes odoo/enterprise#65946

X-original-commit: 53bb670d6fd746388061726046e686a7c0941de1
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Abdessamad Zouiten (abzo) <abzo@odoo.com>