[IMP] account_batch_payment: open o2m payment form view instead of popup
When inside a batch payment form view there is a O2M field `payment_ids`
which when clicked is opening the payment form view inside a dialog.

The issue with that is that since [1], stat buttons were excluded from
dialogs because in some flows, it was creating issues in the user flow,
particularly in cases involving intermediary records that were not yet
saved yet.

But in the batch payments form view, when clicking on a payment, it's
important to see those stats buttons, otherwise it leads to bad UX and
make people waste time.

This commit is making it so there is a "View" button appearing on the
payment lines which will open the form view in the current window,
adding itself to the breadcrumb.

It's possible since commit [2] but sadly for it to work the list need to
be editable, which this commit is thus changing too.

The fields are receiving a `readonly` attribute as we don't want the
payment fields to be editable through the batch payment o2m tree view.
It would be error prone and is not the goal of that view.

[1]: https://github.com/odoo/odoo/commit/473e629a099bfd4b5334951a54df48e37c8612c7
[2]: https://github.com/odoo/odoo/commit/258e6a019a21042bf4f6cf70fcce386d37afd50c

task-3973116

closes odoo/enterprise#65029

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>