[FIX] account: always get tax tags using the base term when they are translatable
To reproduce the issue:

1) Make an invoice using a tax impacting some tag
2) Add a tanslation to that tag
3) Switch the user language to the one you added a translation for
4) Open the tax report: the line the tag is linked to has a value of 0
===> It should contain the value you added in 1)

l10n_multilang makes the 'name' field of account.account.tag translatable. Because of that, it is important to always check the tags matching a report expression's formula using en_US as the language, to make sure the formula of the report expression matches the tag name (since that expression's formula is not translatable).

closes odoo/enterprise#67620

X-original-commit: d1db61a2d579418f8b35f90884dd1bdc3f84372e
Related: odoo/odoo#175001
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>