[FIX] account_followup: reintroduce default value in email_recipient_ids
Steps to reproduce the issue:

1. Go to _Accounting > Customers > Follow-up Reports_ select one of the reports
2. Click on "Follow up"
3. Open the template in the Content Template field
4. Empty the To (Partners) field in this template
5. Save and close
6. Close the Follow Up and reopen it
5. There is no address in the field

Explanation:

With previous fix, we removed the `default_get` value but did not add a default one in the compute.

Suggested fix:

The default value will be the same as the one previously in `default_get`. The purpose is to restore the previous workflow and then add the template's values.

opw-3878125

closes odoo/enterprise#65078

X-original-commit: 011ce344a1ec837143d4fd42eabc5f7a48f1fa03
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>