[FIX] mrp_workorder: force the SN assign only for the final operation
steps to reproduce:

- Create a storable product tracked by SN with a bom with two operations
    in the same workcenter:
    - Operation 1: with at least one component consumed
    - Operation 2
- Create and confirm an MO using that bom
- Go to the shopfloor, validate all the quantities of the operation 1
and mark as done without setting a SN on the final product

> Invalid Operation: You should provide a lot/serial number for the
    final product

Expected behavior:

You should be able to validate the operations of the MO without
assigning an SN but you should not be able to mark the MO as done
without assigning the SN to the final product.

Note: if no component were consumed in an operation, the invalid
operation would not be raised.

Cause of the Issue:

The invalid operation is raised because of these lines:
https://github.com/odoo/enterprise/blob/1b0435a92492d85f28549a207fcdb7224b457c66/mrp_workorder/models/mrp_workorder.py#L433-L434
However, the existence of "self.move_raw_ids" only indicates that a
component has to be consumed in this opration rather than the existence
of a move requiring a the SN of the final product to be set.

opw-3987022

closes odoo/enterprise#66174

X-original-commit: eb80ea5f7a2ac8bdfe4839d814d1394601ad37d3
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>