[IMP] account_accountant: Kanban quick create
This commit adds the quick create feature to the bank reconciliation kanban view. (no more popup)
In order to make bank reconciliation more intuitive and overall easier to use, we are opting for a quick create feature that is similar to the way it currently works in apps like CRM or Project. This makes the reconciliation flow a lot more streamlined.

Task ID: 3916563

closes odoo/enterprise#65891

Related: odoo/odoo#171551
Signed-off-by: Laurent Smet (las) <las@odoo.com>