[IMP] mail: add clickable html tag to activity component
Before this PR, t-on-click was directly on `<i>` and `<span>` elements. This
is not correct since they are not clickable html elements and could be a
challenge for people using accessibility tools.

This PR adds clickable elements to the t-on-clicks

Community PR: https://github.com/odoo/odoo/pull/170000/

closes odoo/enterprise#65053

Related: odoo/odoo#170000
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>