[FIX] sale_subscription: inform that the invoice creation is forced
Issue:
When a transaction is done for a subscription, we force the creation
and sending of an invoice.
However, depending on the localization of the company, many additional
processes can be performed when generating/sending an invoice, like
the generation/sending of an EDI document.
In these cases, we may want to distinguish these forced creations from
the manual ones as the treatment of the invoice could differ depending
on the case.

opw-3936227

closes odoo/enterprise#65086

Related: odoo/odoo#170144
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>