[FIX] stock_barcode: scan packaging and weight
Issue
=====

When a GS1 barcode contains both a packaging and a weight, the quantity
is wrong because we use the weight as the quantity and multiply it by
the packaging quantity.

How to reproduce
================

- Enable "Product Packagings" and "Units of Measure";
- Select GS1 nomenclature as the Barcode Nomenclature;
- Create a product with a packaging for multiple quantity (e.g.: 4
  Units) and with a valid GTIN (e.g.: 12345600012349);
- In the Barcode app, create a new receipt;
- Scan a GS1 barcode containing both the packaging barcode and a weight
  (e.g.: 10123456000123493103001500)
  => The scanned weight is used as the quantity and is multiplied by the
     packaging quantity (in this example, the quantity will be 6,
     because 4 x 1.5 = 6)

Expected Behavior
=================

Since the associated product uses Units as UoM, the scanned weight can't
be converted into quantity (we can't convert g or kg into units), so
this information should be ignored and the packaging quantity should be
used alone instead.

Note that the packaging quantity will still multiply the scanned
quantity if the share the same UoM category. This behavior is still
supported to be able to scan multiple packagings at once.
That said, this use case make sense for products using Units only.

OPW-3988826

closes odoo/enterprise#66403

X-original-commit: a0ab3b439219953e8e2f0f981c71eb03efa7b8a0
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>