[REF] web_tour: remove unused functions
The tourError function was removed because it duplicates console.error()
which throws exactly the same kind of error on the runbot.
The getDifferentParents function was removed from tour_utils as it is
no longer used.
The triggerPointerEvent function has been removed from tour_utils as it
is no longer used.

task~3974087

closes odoo/enterprise#65090

Related: odoo/odoo#170158
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>