[FIX] documents{_sign}: fix access error applying sign action
How to reproduce:
- Install documents_sign and log as admin
- In documents, create an action (Workspace: Finance, Create: PDF to Sign)
- Then upload a pdf in the workspace Finance
- Log as Marc Demo
- In documents, select the PDF uploaded as admin
- Click on the action created

You get the error: "Sorry, you are not allowed to access this document."
while the expected behavior is that users who have access to the document and
to Sign should be able to use the action. We solve that in this commit.

We also add a test checking that applying an action to sign a pdf document not
owned by the user that perform the action, doesn't duplicate the attachment if
the attachment is owned by the document (res_id and res_model pointing to the
document; the ownership is transferred to the sign.template created) and
duplicate it otherwise.

Technical note: the error was caused by the trick used in the override of
sign_template.create in documents_sign which was setting the res_model to False
and res_id to 0 for attachment linked to document to allow the super method to
link that attachment to the created sign_template (by releasing the ownership
of the attachment from the document). But as ir_attachment prevents to link an
attachment with res_id = 0 to a record if it is not done by the owner of that
attachment, there was an error. We remove that tricks by defining a method that
determine if sign_template can take the ownership of the attachment or not
which we override in document_sign and make it return True when the attachment
is owned by a document.

Task-4000988

closes odoo/enterprise#73338

X-original-commit: a0d290a95226f80c9993644d871c8fe7020afb8a
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>