[IMP] website_knowledge: add Open Graph tags for published articles
When a user publishes an article and shares the access link with a friend
on Facebook, WhatsApp, or other social media platforms, most platforms
fetch the website's Open Graph metadata to display a preview of the link.

For the website's knowledge module, we have not defined any Open Graph
metadata. The system currently uses default values loaded from the
`ir.ui.view` model and the preview does not give any information about
the article.

To enhance the link preview on the external platforms and enhance the
SEO of the articles, we will manually set the Open Graph metadata using
the information about the article. This should make people more inclined
to click on the link.

task-4001922

closes odoo/enterprise#65108

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>