[FIX] web: remove unnecessary variable override
This commit removes an unnecessary variable override within our `.dark`
files.

Before `saas-17.3`. which introduces a migration to Bootstrap v5.3,
the `$card-bg` was set to `white` in light mode, which needed an override
in dark mode as it'd still be pure white.

As we change that value to `$o-view-background-color` which matches the
`$card-cap-bg` value, we can remove that override.

task-4004392
part of task-3883628

closes odoo/enterprise#65471

X-original-commit: c93bc73b8139c313c932267e3ab01bbdc98b8413
Related: odoo/odoo#170796
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>