[FIX] sign: fix button contrast
This commit fixes a contrast issue on a primary button within the sign
module due to a `text-white` class applied.

The default button utility classes already provide a good contrast,
meaning that adding extra class is not useful. To fix the issue,
we simply remove the class from the button.

task-3692048

closes odoo/enterprise#65414

X-original-commit: 4a8cca2186c5f3825aca9bb9ee5b5e79edc84317
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>