[FIX] sign: fix button contrast
This commit fixes a contrast issue on a primary button within the sign
module due to a `text-white` class applied.

The default button utility classes already provide a good contrast,
meaning that adding extra class is not useful. To fix the issue,
we simply remove the class from the button.

task-3692048

closes odoo/enterprise#64801

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>