[FIX] l10n_lu_reports: fix a test violating the tax lock date
Currently the tax lock date is not checked correctly in a test
when posting a move.

This problem is solved in the related community PR.

This commit ensures the test does not fail after the community fix.
To do this we just remove and re-introduce the tax lock date.

closes odoo/enterprise#65374

X-original-commit: b36a0924ed137504e8b98822ea68bd011befd18c
Related: odoo/odoo#170651
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>