[FIX] l10n_lu_reports: fix a test violating the tax lock date
Currently the tax lock date is not checked correctly in a test
when posting a move.

This problem is solved in the related community PR.

This commit ensures the test does not fail after the community fix.
To do this we just remove and re-introduce the tax lock date.

closes odoo/enterprise#65262

Related: odoo/odoo#170475
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>