[FIX] account_reports,*: don't show buttons on prefix group lines
*l10n_account_customer_statements,l10n_my_reports,l10n_us_reports

Currently some reports add buttons on their lines to perform some action
on them. However, when prefix groups are used, the buttons are also
shown on these lines, but refer to nothing and will thus either cause an
error when clicked or lead to a new record.

This commit fixes that by providing a generic hook in the `line_name`
template to inject buttons. This generic hook checks whether the line is
a prefix group line and doesn't render the buttons if it is.

All other reports were adapted to use this hook now.

task-3932945

closes odoo/enterprise#66368

X-original-commit: ff203e2e4802cfd5c135d0bc4020d987ecd73b80
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>