[IMP] mail, *: replace "style-inline" by HtmlMailField
The purpose of this commit is to remove the notion of "inline" in html_editor.
In the old html_field (web_editor), you could use a "style-inline" option
to inline the html save. This feature is only used to generate html
used in emails. We therefore decided to create the HtmlMailField, which
will take care of html fields that need to have their content inlined.

This allows us to:
- simplify the html_field by moving the inline logic to /mail
- remove the "style-inline" option

Part-of: odoo/enterprise#64425
Related: odoo/odoo#149028
Signed-off-by: Géry Debongnie <ged@odoo.com>