[IMP] web_studio: add columns in qweb tables
The report editor edits raw qweb. In this setting, tables can contain illegal nodes with regard to html spec.
So we need to convert those tables into something else, that is, as simple div-like elements displayed as flex.

Before this commit, it was not possible to add/remove columns.
This commit proposes a solution to this.
Note that, given the use cases, removing/adding rows is still impossible because we assume tables of that type
are doing a `t-foreach` in the row dimension.

task-3940106

Part-of: odoo/enterprise#67088
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>