[FIX] planning: change display condition of buttons in planning gantt
With task-3336242, the control panel's dropdown button's display
condition was changed from <= xl display to <= md display. Hence, this
PR changes the display condition of buttons in the planning gantt
view to align with this new behavior.

The bootstrap classes that were used to attain this effect are now
removed, replaced by a t-if="env.isSmall". This allows us to have only
one of the two implementation present in the dom, and to change the
display condition to <= md.

task-4005521

closes odoo/enterprise#66941

X-original-commit: a4b8e8d5ae98190d13c27cf02fa38dba081b0dcc
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>