[FIX] l10n_pe_edi: fix display of QR-code on invoice
Steps to reproduce:
- Install Accounting and l10n_pe_edi
- Switch to a Peruvian company (e.g. PE Company)
- In Accounting settings, select "SUNAT" as Signature Provider
and activate "QR Codes" option
- Create an invoice for a Peruvian contact
- Sent the invoice to Peru UBL 2.1
- Generate the invoice via "Send & Print" button

Issue:
The QR code doesn't appear on the invoice.

Cause:
There's a condition on the parent element of the QR element that checks
if "spot" is defined.

Solution:
The display of the QR code should be independent of the "spot" variable.

opw-3990721

closes odoo/enterprise#65589

X-original-commit: 0fd6be2d05a8f45448355b637072d917e0d2ef10
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>