[IMP] account_accountant{_fleet}: bank rec UI improvements
This commit aims to:
- make more coherent information displayed in the list view
and the kanban view of the bank reconciliation widget.
To do so we now display `ref`, `narration`, `bank_account` `foreign_currency` and `amount_currency`
within the "Manual Operations" tab.
- display information even after the statement line is reconciled.
To do so we made the "Manual Operations" tab visible even after validation, with its information
in readonly.
- make the cog menu available even if the transaction if validated
- fix some labels of readonly fields that were not greyed-out

task-3996739

closes odoo/enterprise#65313

Signed-off-by: Laurent Smet (las) <las@odoo.com>