[FIX] account_reports,*: apply filter on clicking and not on close
It's really unintuitive and non-Odoo to have to close the filter box to be able to apply a filter on reports.
Let's change that.
We now let user click on filters while it's loading.
We then delay filters that make a rpc call by 500 ms.
The goal being that we don't do 5 calls if the user wants
to select 5 journals.

It implies several difficulties, including a reloading of the
data or the options that were done when clicking.
For that, we use incrementCallNumber (to know if the options
sent by the python are the ones needed to be reloaded).

task-3912381

closes odoo/enterprise#68330

X-original-commit: 10d61ca19ea850be7023043051c31f2b4bd3354c
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>