[IMP] web_studio,*: adapt code to new list column widths
PR odoo/odoo#170511 reworks the column widths logic of the list
view. Code had to be adapted accordingly.

In studio, the table was too wide, so there was always an horizontal
scrollbar. It was due to two things: the column hooks, which were
supposed to have a 1px width, but which were actually 4px because
of their padding, the padding of the editor which wasn't took into
account when computing the available space. We fixed that issue
by setting the padding on the o_content instead of the table.

In Appraisal, widths are hardcoded in the arch anyway, so the js
override seemed useless.

Task 3866837

Part-of: odoo/enterprise#65329
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>