[FIX] sale_subscription: double check on invoices
https://github.com/odoo/enterprise/commit/f510c4098cc4696dec3fbc7d6cd3164c6fb1d6bd
prevented to close a subscription if there was no invoice.
Unfortulately, sale_order.order_line.invoice_lines.move_id is not always
a reliabable way to check for invoices. Sme flows could replaces all the
sale.order.line and therefore the link to the invoices would be lost.

In that case, for subscription orders, using the subscription_id field
on the account.move.line is more reliable. We use both values to be
sure.

X-original-commit: 3dffd42b5d651299595349f409e478e9e9da54a4
Part-of: odoo/enterprise#65478
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>