[FIX] l10n_es_reports: show the NIF number of non EU partners in tax report
Issue:
The VAT lines in tax reports do not populate correctly.

Steps to reproduce:
- Activate l10n_es module
- Go to Accounting > Reporting > Statement Reports > Tax Report
- Select a 'Generic Tax Report' and click on 'VAT Record Books (XLSX)'
- In the report only Spain and partners without countries have a NIF number

Cause:
A line of code is only populating the NIF field in the report if the partner has no country or if the country is Spain.

In addition, other lines related to the NIF are not populated as they should.

Solution:
Add conditions to populate the NIF lines as the specs demands it.

opw-3918623

closes odoo/enterprise#63923

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>