[FIX] l10n_pe_edi_stock: Dont clear ticket number if connection error
Currently, if there is a connection error while retrieving the CDR, we
clear the ticket number. This will cause the user to send a duplicate
delivery guide if he clicks on 'Retry', which is not what we want.

Instead, we should keep the ticket number in all error cases except
explicit rejection from SUNAT, so that a retry will try to retrieve
the CDR again using the existing ticket number.

opw-3757843

closes odoo/enterprise#65580

X-original-commit: 748a1413cb5e6e60101e62547090f1b265765677
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>