[FIX] sendcloud_service: use customer address when service point used
Before this commit:

When the user chooses a pickup point, the res.partner of the sale order is replaced with
data from the pickup point, while the previously set customer's data is set as the partner's parent.
And Sendcloud relies on a service_point_id field in its API rather than a raw address.

Which causes double printing of the service point address on the label.
Moreover, Sendcloud explicitly requires the customer's (and not the pickup point) address in its API.

After this commit:

Both the service_point_id and the customer's delivery address are sent to Sendcloud.
Both addresses are printed on the label correctly.

opw-3977565

closes odoo/enterprise#66235

X-original-commit: 45767074b1840b066427478b892cf136901d032c
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>