[FIX] pos_preparation_display:Allow order to appear on multiple displays
Description:
The current setup does not allow orders to appear on multiple preparation
displays. This happens as the stage is created for the first display
opened and the order is just linked to that stage as the stageless orders
function would not return it for other displays. This is not the intended
behavior and orders need to appear in multiple displays.

Steps to reproduce the issue:

- create a pos and 2 preparation displays for it

- create an order in pos

- check the order in both preparation displays

- The order only appears in one

Source of the issue:
The issue is introduced as the _get_stageless_orders_in_display
would not return orders that has a stage for any other display.

The fix is done by adding an extra leaf to check that the stages
being checked against are relevant to the current display.

closes odoo/enterprise#64669

Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>