[IMP] industry_fsm_sale: prevent empty sol creation for timesheet
With this commit, we prevent the creation of an empty sol for timesheets when
products are added to the task but no timesheet entries are made. and, the task
is marked as done.

test case is added to ensure that when only a product is linked to a task and no
timesheet is added, the SOL is created only for the product and not for the
timesheet.

task-3995703

closes odoo/enterprise#65750

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>