[FIX] web_enterprise: use back button don't break filters in search view
Open a kanban view or a list view;
Enter a filter in the search view;
Open a record;
Go back to the multi-record view using the browser (back button).

Before this commit, the previously entered filter is not there anymore.
Now, the filter stays. To achieve this, before leaving the multi-record
view, the global state is saved on the history state.

X-original-commit: ac5d309f3ee91db17e738fe2da9104810911e3d4
Part-of: odoo/enterprise#65468
Related: odoo/odoo#170792
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>