[REF] spreadsheet_edition,*: generalize spreadsheet dialog selector
*spreadsheet_edition, documents_spreadsheet, spreadsheet_dahboard_edition

TLDR: no xpath, no component inheritance, only a small python method.

This commit prepares task 3761590 in which a new spreadsheet model is
introduced.
Users will be able to add pivots/lists in those spreadsheets.

Currently, the dialog to select a spreadsheet when inserting a view is not
entirely generic.
For each model, the xml template needs to be patched, and a new component
created.

With this commit, the dialog no longer requires any front-end change. Everything
is python-driven.

New spreadsheet models only need to implement method `_get_spreadsheet_selector`
to appear in the selection dialog.

Task: 4011300
Part-of: odoo/enterprise#65397
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>