[IMP] *: remove generate offer wizard and update offer views
*: hr_contract_salary, l10n_be_hr_contract_salary, test_l10n_be_hr_payroll_account

The information in the `generate offer wizard` is the same as the information in
the offer form view (in which you can also edit the offer similar to the wizard)
that is shown after the offer is generated. Thus, the wizard doesn't provide any value.
Also, Studio is disallowed to add fields on wizard. So it's a hindrance to the flow.

1. These changes remove the `generate offer wizard`, move any warning that was generated
on the wizard to the offer form view and make `Generate Offer` button lead directly to that form view.

2. The changes also update the offer views:
    1. In offer form view:
        1. 'Title' is now editable and the default title doesn't include the offer creation date or the employer budget.
        2. `Contract Template`, `Employee Job` and `Department` are now editable.
        3. `Refual Reason` and `Refusal Date` now only appear if they aren't empty.
    2. In offer list view:
        1. A new field `created by` is added to the list view and choose the creator of the offer.
        2. The fields are reordered.

3. In addition to that, the changes adapt tours to the new flow after removing the salary wizard.

task-4008437

Part-of: odoo/enterprise#65529
Related: odoo/upgrade#6221
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>