[FIX] whatsapp: new message separator not shown
When sending a lot of messages, it can happen that the new message
separator does not appear anymore.

The community counterpart of this PR fixes this issue but modify a
notification payload that is asserted in a whatsapp test.

This PR updates this test to match the notification content.

closes odoo/enterprise#65394

Related: odoo/odoo#170601
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>