[FIX] project_enterprise: avoid overwriting handpicked planned dates
Versions
--------
- 15.0+

Steps
-----
1. Have a task assigned to you;
2. ensure planned hours are empty;
3. plan the task partially outside working hours (e.g. 16:00 -> 21:00);
4. save.

Issue
-----
Planned hours are overwritten.

Cause
-----
Commit 52536bf8bb5 added logic to provide schedule-based start dates for
tasks created in Gantt view, as well as other tasks, by overwriting the
provided (default) date values.

Solution
--------
As of commit 0eda2ab1332, Gantt task dates are planned in JS, and
skipped in the back-end due to the `smart_task_scheduling` context
value. Therefore we can skip the overwriting of dates if we're working
on a single record, which would usually be from a form view. This way,
we can still do schedule-based planning when batch processing from list
view.

opw-3947876

closes odoo/enterprise#75454

X-original-commit: 1ac3a830b1ddc25bb848b606b9622fae92e1fcf5
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>