[FIX] account_online_synchronization: call consent expiring date after success
The aim of this commit is reverting a part of this commit [1].
We have to do that because by removing the call to consent expiring
date after the sucess, users don't receive the new consent date
directly. Users have to wait until the next call to the refresh token and
this call is made every 24h so users have the feeling that the consent wasn't
updated.
In one case, we don't have the consent expiring date provided by the TPP
directly. For that case, we keep the call after the renew token.

[1]: https://github.com/odoo/enterprise/commit/25ec0ba479372126f84fd1ece6f34aea7b1eb829

closes odoo/enterprise#67019

X-original-commit: e073dad49817451b5330b5e342ab172a326cede9
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>