[FIX] account_reports: hide lines with string values
Some reports display some informational columns (like the VAT for a
partner, or the country of origin). These columns sometimes prevent
the `Hide Lines at 0` option to correctly do its work: the desired
behavior of this option is to hide the lines where the amounts are null.
It shouldn't care about the "informational" columns, that are usually
always present

This commit force the `Hide Lines at 0` feature to ignore columns
displaying information other than float, integer of monetary ones.

task-4006235

closes odoo/enterprise#65622

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>