[FIX] website_sale_subscription : Update price of recurring products
Steps to reproduce:
	1. Navigate to the 'Recurring Plan' page and create a plan for the annual option.
	2. Under a company setting:
		- Create a new product named test.
		- Add attributes/variants such as v1, v2.
		- In the 'Recurring Plan' tab, assign v1 and v2 to the annual plan with distinct
		prices(e.g., v1 = , v2 = ).
		- Ensure no product category is added (or leave it by default).
	3. Go to 'Website':
		- Locate the product and switch between variants v1 and v2.
		- Observe that the product price remains unchanged upon switching, but if you
		add the product to the cart, the prices differ.

Current behavior before PR:
The recurring product's price doesn't get updated when choosing another
variant. This is happening because the xml template shown in this form
is static and not getting re-rendered when a value changes.
https://github.com/odoo/enterprise/blob/saas-17.2/website_sale_subscription/views/templates.xml#L25:L42

Desired behavior after PR is merged:
We are updating the data shown in the xml tag responsible to show the
price from JS side by writing the new value in the textContent value of this
xml tag.

opw-3983826

closes odoo/enterprise#65919

X-original-commit: 7d96b65f3755cf174cd040a4ddd43b947909d8f3
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>