[IMP] web_enterprise: better 'refresh sub. status' btn visibility on exp. panel
  For on premise instances, when the database expire and user renew
its subscription, the update is not immediate; on the expiration panel
there is a already a button to forcefully refresh the subscription
status and unblock the database, but it's not really visible (its not
clear that the 'refresh' icon right next to "Renew your subscription" /
"Upgrade your subscription" buttons is just to refresh the subscription
status and not to re-do that action again).

  This commit now shows the 'Refresh subscription status' action as a
textual button to make it more visible to users.

opw-3942907

closes odoo/enterprise#70387

X-original-commit: 0f0b07f1537db8ab7748ef05290489d40c75a823
Signed-off-by: Xavier Alt (xal) <xal@odoo.com>