[FIX] l10n_pe_edi: Don't retrieve CDR if invoice with same name was sent
In Peru, we don't prevent invoices with the same name from being
created. Invoices with the same name, same document type and same
company RUC will be reported to SUNAT with the same `edi_filename`.

SUNAT rejects those invoices as duplicates, which is good. However, our
fallback mechanism of retrieving the existing CDR for those invoices
and marking the invoice as sent is not good in this situation. Indeed,
the user should be notified that the invoice was rejected, prompting
them to resequence their invoice. Instead, they get the impression that
their invoice was correctly sent to SUNAT.

To fix this, we don't try to retrieve an existing CDR if an invoice was
rejected as duplicate, but there already exists an invoice with the same
edi_filename that was sent to SUNAT.

opw-3900393

closes odoo/enterprise#67161

X-original-commit: 3d835d90fba9bf7092df3eb0fe065270993a334f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>