[FIX] account_report: traceback when no name
Before this commit, when activating analytic the total column of horizontal
group was displayed which was not the intended behavior but also a traceback
appeared when selecting a None horizontal group because the name was undefined.

Also changing the fact that the horizontal group should reload the page when
changing the filter.

closes odoo/enterprise#66159

Task: 4013957
X-original-commit: d7ee328f4fdd3efffc34be9d9b20cfe0c658cc98
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>