[ADD] l10n_it_pos: add support for Italy fiscal printer
Purpose
Since July 2019, it's mandatory in Italy to have your POS connected to a "fiscal printer", meaning a printer connected to the government. so the goal is to make POS compliant with Italian fiscal law integrating it specifically with the Epson Italian Fiscal Printers.

Updates Introduced:
- Support the Epson printer following operations:
    - print Fiscal Receipt
    - print Fiscal Invoice
    - print X Report
    - print Z Report
    - print XZ Report
    - re-print Receipt
    - refunds
    - discounts

- Adapt the POS tax calculations to be consistent with printer calculations

task-2092597

closes odoo/enterprise#70164

Related: odoo/odoo#180458
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>