[FIX] l10n_mx_edi_website_sale: add French translations
This module adds user facing elements to ecommerce (i.e. the Website
> Shop). This results in part of the checkout process showing up in
English when customers are using a language other than Spanish. We make
an exception for translating a localization into a language other than
the one spoken in the country it's specific to because this is customer
facing (i.e. in the website) and we do not want to restrict the
language of the potential ecommerce customers.

Unfortunately some of the terms that show in the extra step added by
this model are defined in other `l10n_mx_*` modules and since we don't
want to have to track down only these specific terms or translate all
of the Mexico localization into French, we leave some parts
untranslated (specifically the invoicing specifics)

opw-3881171

closes odoo/enterprise#65563

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>